Move component PositiveNumberByDiagnosedDateCard from 'components/cards' to 'components/index/CardsReference'
